Refactor: Push resource list enrichment into ResourcesRepository
Moves the privacy filtering, ratings lookup, tags lookup, and the
mapping logic that creates `ResourceListModel` objects out of
`ResourcesViewModel` and into a new `getEnrichedLibraryListModels`
method in `ResourcesRepository`. This keeps the ViewModel focused
on screen state and sync logic rather than cross-feature joins.
